Move GetEnvBuildURL to azdo package
This is where it belongs: it's specific to CI builds in AzDO. Moving it removes clutter from the releasego package and it may be used by other packages in the future.
